requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django>=4.2.0
django-cors-headers>=4.4.0
django-debug-toolbar==5.2.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
joblib==1.5.1
nltk==3.9.1
openai==1.91.0
# spacy>=3.7.0  # 현재 사용 안함
packaging==25.0
pillow==11.2.1
pip==23.0.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
# pytesseract==0.3.13  # OCR 사용 안함
python-dotenv>=1.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
setuptools==65.5.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0

# postgresql
psycopg2-binary>=2.9.9
gunicorn>=21.2.0

# Google OAuth2
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0